libwimaxll: fix wimaxll_recv() mismatched return value declaration and printf format
Fix to bug http://linuxwimax.org/bugzilla/show_bug.cgi?id=2 reported
by Paul Donohue.

Undetected mistmatches in wimaxll_recv():

- declared and defined return types
- print format

caught by a 64-bit build.

+++ include/wimaxll.h
@@ -309,7 +309,7 @@ const char *wimaxll_ifname(const struct wimaxll_handle *);
 
 /* Wait for data from the kernel, execute callbacks */
 int wimaxll_recv_fd(struct wimaxll_handle *);
-int wimaxll_recv(struct wimaxll_handle *);
+ssize_t wimaxll_recv(struct wimaxll_handle *);
 
 /* Default (bidirectional) message pipe from the kernel */
 ssize_t wimaxll_msg_write(struct wimaxll_handle *, const char *,
